exeptions add

+++ src/main/java/matveyodintsov/scoreboard/repository/PersistenceRepository.java
@@ -1,5 +1,6 @@
 package matveyodintsov.scoreboard.repository;
 
+import matveyodintsov.scoreboard.util.AppConst;
 import matveyodintsov.scoreboard.util.HibernateUtil;
 import org.hibernate.Session;
 import org.hibernate.query.Query;
@@ -17,23 +18,29 @@ public abstract class PersistenceRepository<T> implements Repository<T> {
 
     @Override
     public void save(T object) {
+        if (object == null) {
+            throw new IllegalArgumentException(AppConst.Message.CANNOT_SAVE);
+        }
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             session.beginTransaction();
             session.saveOrUpdate(object);
             session.getTransaction().commit();
         } catch (Exception e) {
-            e.printStackTrace();
+            throw new RuntimeException(AppConst.Message.CANNOT_SAVE);
         }
     }
 
     @Override
     public void delete(T object) {
+        if (object == null) {
+            throw new IllegalArgumentException(AppConst.Message.FAILED_DELETE);
+        }
         try (Session session = HibernateUtil.getSessionFactory().openSession()) {
             session.beginTransaction();
             session.delete(object);
             session.getTransaction().commit();
         } catch (Exception e) {
-            e.printStackTrace();
+            throw new RuntimeException(AppConst.Message.FAILED_DELETE);
         }
     }
